# More efficient `PoissonBinomial`