--- a/test/poissonbinomial.jl
+++ b/test/poissonbinomial.jl
@@ -29,6 +29,7 @@ naive_sol = naive_pb(p)
 
 @test Distributions.poissonbinomial_pdf_fft(p) ≈ naive_sol
 @test Distributions.poissonbinomial_pdf(p) ≈ naive_sol
+@test Distributions.poissonbinomial_pdf(Tuple(p)) ≈ naive_sol
 
 @test Distributions.poissonbinomial_pdf_fft(p) ≈ Distributions.poissonbinomial_pdf(p)
 
@@ -46,23 +47,23 @@ for (p, n) in [(0.8, 6), (0.5, 10), (0.04, 20)]
     @test maximum(d) == n
     @test extrema(d) == (0, n)
     @test ntrials(d) == n
-    @test entropy(d)  ≈ entropy(dref)
-    @test median(d)   ≈ median(dref)
-    @test mean(d)     ≈ mean(dref)
-    @test var(d)      ≈ var(dref)
-    @test kurtosis(d) ≈ kurtosis(dref)
-    @test skewness(d) ≈ skewness(dref)
+    @test @inferred(entropy(d))  ≈ entropy(dref)
+    @test @inferred(median(d))   ≈ median(dref)
+    @test @inferred(mean(d))     ≈ mean(dref)
+    @test @inferred(var(d))      ≈ var(dref)
+    @test @inferred(kurtosis(d)) ≈ kurtosis(dref)
+    @test @inferred(skewness(d)) ≈ skewness(dref)
 
     for t=0:5
-        @test mgf(d, t) ≈ mgf(dref, t)
-        @test cf(d, t)  ≈ cf(dref, t)
+        @test @inferred(mgf(d, t)) ≈ mgf(dref, t)
+        @test @inferred(cf(d, t))  ≈ cf(dref, t)
     end
     for i=0.1:0.1:.9
-        @test quantile(d, i) ≈ quantile(dref, i)
+        @test @inferred(quantile(d, i)) ≈ quantile(dref, i)
     end
     for i=0:n
-        @test isapprox(cdf(d, i), cdf(dref, i), atol=1e-15)
-        @test isapprox(pdf(d, i), pdf(dref, i), atol=1e-15)
+        @test isapprox(@inferred(cdf(d, i)), cdf(dref, i), atol=1e-15)
+        @test isapprox(@inferred(pdf(d, i)), pdf(dref, i), atol=1e-15)
     end
 
 end
@@ -88,11 +89,11 @@ for (n₁, n₂, n₃, p₁, p₂, p₃) in [(10, 10, 10, 0.1, 0.5, 0.9),
     pmf2 = pdf.(b2, support(b2))
     pmf3 = pdf.(b3, support(b3))
 
-    @test mean(d) ≈ (mean(b1) + mean(b2) + mean(b3))
-    @test var(d)  ≈ (var(b1) + var(b2) + var(b3))
+    @test @inferred(mean(d)) ≈ (mean(b1) + mean(b2) + mean(b3))
+    @test @inferred(var(d))  ≈ (var(b1) + var(b2) + var(b3))
     for t=0:5
-        @test mgf(d, t) ≈ (mgf(b1, t) * mgf(b2, t) * mgf(b3, t))
-        @test cf(d, t)  ≈ (cf(b1, t) * cf(b2, t) * cf(b3, t))
+        @test @inferred(mgf(d, t)) ≈ (mgf(b1, t) * mgf(b2, t) * mgf(b3, t))
+        @test @inferred(cf(d, t))  ≈ (cf(b1, t) * cf(b2, t) * cf(b3, t))
     end
 
     for k=0:n
@@ -104,7 +105,7 @@ for (n₁, n₂, n₃, p₁, p₂, p₃) in [(10, 10, 10, 0.1, 0.5, 0.9),
             end
             m += pmf1[i+1] * mc
         end
-        @test isapprox(pdf(d, k), m, atol=1e-15)
+        @test isapprox(@inferred(pdf(d, k)), m, atol=1e-15)
     end
 end