@devmotion 2021-02-23
I compared the implementation in this PR with an implementation of an immutable struct that uses `pmf::Vector{T}` which is empty in the uninitialized state and resized and then filled with an in-place variant of `poissonbinomial_pdf` if needed.

# Without `sizehint!`

Benchmarks without `sizehint!(pmf, length(p) + 1)` in the constructor:

```julia
julia> using Distributions, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p = rand(100);

julia> @btime PoissonBinomial($p);
  110.461 ns (1 allocation: 80 bytes)

julia> @btime PoissonBinomial($p; check_args=false);
  31.534 ns (1 allocation: 80 bytes)

julia> @btime mean(PoissonBinomial($p));
  121.792 ns (1 allocation: 80 bytes)

julia> @btime var(PoissonBinomial($p));
  217.599 ns (1 allocation: 80 bytes)

julia> @btime mode(PoissonBinomial($p));
  4.675 μs (4 allocations: 960 bytes)

julia> @btime median(PoissonBinomial($p));
  4.706 μs (5 allocations: 1.81 KiB)

julia> @btime modes(PoissonBinomial($p));
  5.808 μs (52 allocations: 6.20 KiB)

julia> @btime mgf(PoissonBinomial($p), 0.1);
  152.058 ns (1 allocation: 80 bytes)

julia> @btime cf(PoissonBinomial($p), 0.1);
  360.134 ns (1 allocation: 80 bytes)

julia> @btime entropy(PoissonBinomial($p));
  5.155 μs (4 allocations: 960 bytes)

julia> @btime mgf($(PoissonBinomial(p)), 0.1);
  42.884 ns (0 allocations: 0 bytes)

julia> @btime cf($(PoissonBinomial(p)), 0.1);
  246.104 ns (0 allocations: 0 bytes)
```

# With `sizehint!`

With `sizehint!(pmf, length(p) + 1)` in the constructor I get:

```julia
julia> using Distributions, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p = rand(100);

julia> @btime PoissonBinomial($p);
  139.443 ns (2 allocations: 896 bytes)

julia> @btime PoissonBinomial($p; check_args=false);
  57.826 ns (2 allocations: 896 bytes)

julia> @btime mean(PoissonBinomial($p));
  146.683 ns (2 allocations: 896 bytes)

julia> @btime var(PoissonBinomial($p));
  240.535 ns (2 allocations: 896 bytes)

julia> @btime mode(PoissonBinomial($p));
  4.688 μs (4 allocations: 960 bytes)

julia> @btime median(PoissonBinomial($p));
  4.714 μs (5 allocations: 1.81 KiB)

julia> @btime modes(PoissonBinomial($p));
  5.825 μs (52 allocations: 6.20 KiB)

julia> @btime mgf(PoissonBinomial($p), 0.1);
  175.456 ns (2 allocations: 896 bytes)

julia> @btime cf(PoissonBinomial($p), 0.1);
  385.257 ns (2 allocations: 896 bytes)

julia> @btime entropy(PoissonBinomial($p));
  5.162 μs (4 allocations: 960 bytes)

julia> @btime mgf($(PoissonBinomial(p)), 0.1);
  42.805 ns (0 allocations: 0 bytes)

julia> @btime cf($(PoissonBinomial(p)), 0.1);
  244.248 ns (0 allocations: 0 bytes)
```

# Summary

So in both cases the number of allocations is larger than in this PR and the functions that do not require the evaluation of the probability mass function are slower (which I tried to optimize in this PR). I did not expect these results - what's the explanation for them? Are the compiler optimizations for `Union{Nothing,T}` just "too" good?