@andreasnoack 2021-02-23
Thanks for benchmarking this. The compiler optimizations for the small union here might be really good. Why would e.g. the `mean` function allocate in the latter case but not in the `Union` case?