@devmotion 2021-02-23
> Why would e.g. the mean function allocate in the latter case but not in the Union case?

Yes, IMO this was the most surprising result. I assume that we just see the allocation of the empty vector (80 bytes) in the benchmark without `sizehint!` (with `sizehint!` it increases to 896 bytes) whereas in the `Union` case no vectors have to be allocated. Therefore the number of allocations is reduced and it is faster.