@andreasnoack 2021-02-24
This is puzzling to me
```julia
julia> @btime PoissonBinomial($p);
  90.943 ns (1 allocation: 32 bytes)
...
julia> @btime mean(PoissonBinomial($p));
  104.703 ns (0 allocations: 0 bytes)
```
I'd think that the latter would time construction as well as the mean calculation whereas the first would only time the construction. Why is the first allocating when the second one isn't?