@devmotion 2021-02-24
Yes, this was surprising to me as well. One can get rid of the allocation by changing the returned value:
```julia
julia> f(p) = (PoissonBinomial(p); return)

julia> @btime f($p);
  88.468 ns (0 allocations: 0 bytes)
```
I just don't fully understand why one gets this additional allocation. I inspected the output of `@code_llvm` for both cases, and it seems the main/only difference apart from the returned value are the additional statements
```
; │││ @ /home/david/.julia/dev/Distributions/src/univariate/discrete/poissonbinomial.jl:32 within `_#64'
     %44 = bitcast %jl_value_t*** %ptls to i8*
     %45 = call noalias nonnull %jl_value_t* @jl_gc_pool_alloc(i8* %44, i32 1424, i32 32) #2
     %46 = bitcast %jl_value_t* %45 to %jl_value_t**
     %47 = getelementptr %jl_value_t*, %jl_value_t** %46, i64 -1
     store %jl_value_t* inttoptr (i64 140610215652240 to %jl_value_t*), %jl_value_t** %47
     %48 = bitcast %jl_value_t* %45 to %jl_value_t**
     store %jl_value_t* %13, %jl_value_t** %48, align 8
     %49 = bitcast %jl_value_t* %45 to i8*
     %50 = getelementptr inbounds i8, i8* %49, i64 8
     %51 = bitcast i8* %50 to %jl_value_t**
     store %jl_value_t* inttoptr (i64 140610869056864 to %jl_value_t*), %jl_value_t** %51, align 8
```
in the version with one allocation. So I assume in the non-allocating case (and also the `mean` etc. examples) the `PoissonBinomial` instance is never instantiated but the call of `new{T,typeof(p)}(p, nothing)` is removed by the compiler?