@andreasnoack 2021-03-26
I think we should just stick with your current implementation since the alternative one didn't show improvements.